backend/requirements.txt
# Quiet Eyes Backend Dependencies
# ================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.4
pydantic-settings==2.7.1

# AI Integration
openai==1.59.5

# Database (Supabase)
supabase==2.11.0

# HTTP Client
httpx==0.28.1

# Environment Variables
python-dotenv==1.0.1

# WhatsApp (Twilio)
twilio>=9.0.0

# PDF Generation
reportlab>=4.0.0
python-bidi>=0.6.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
